small refactor splitting things out of codegen-x86
